+++ docs/changelog.md
@@ -1,5 +1,9 @@
 ## CHANGELOG
 
+### UNRELEASED
+
+ - Improve post routing middleware
+
 ### 0.16.1 - 2020-04-10
 
  - Change docs to dark theme